requirements.txt
Flask
python-dotenv
twilio
requests
langchain-google-genai
fastapi
uvicorn
crewai
